Set entire sample to missing CN if cnvnator fails
This happens on a chromosome or sample basis.
See https://github.com/abyzovlab/CNVnator/issues/83

Previously we simply died which results in
truncated files and headaches. At this point in time
failing the entire sample seems wise as we think
this an indication that the sample is of poor
quality.

Closes #227